Conflicts:
	htdocs/bimpcore/objects/Bimp_Product.class.php
	htdocs/bimpequipment/objects/BE_Package.class.php
	htdocs/bimpequipment/objects/Equipment.class.php
	htdocs/bimplogistique/objects/Inventory2.class.php

Former-commit-id: 5138adf0d121d023c6805f14e7e249e4a12d8787

+++ htdocs/bimpcommercial/core/triggers/interface_99_modBimpcommercial_PostProcess.class.php
@@ -9,41 +9,50 @@ class InterfacePostProcess extends BimpCommTriggers
     {
         global $conf;
         $errors = array();
+        $warnings = array();
 
         $object_name = '';
         $action_name = '';
 
-        $bimpObject = $this->getBimpCommObject($action, $object, $object_name, $action_name);
+        $bimpObject = $this->getBimpCommObject($action, $object, $object_name, $action_name, $warnings);
 
         if (BimpObject::objectLoaded($bimpObject)) {
-            $warnings = array();
-            
-            switch ($action_name) {
-                case 'CREATE':
-                    $errors = $bimpObject->onCreate($warnings);
-                    break;
+            $obj_errors = BimpTools::getDolEventsMsgs(array('errors'), false);
+            if (empty($obj_errors)) {
+                switch ($action_name) {
+                    case 'CREATE':
+                        if(method_exists($bimpObject, 'onCreate'))
+                            $errors = $bimpObject->onCreate($warnings);
+                        break;
 
-                case 'VALIDATE':
-                    $errors = $bimpObject->onValidate($warnings);
-                    break;
+                    case 'VALIDATE':
+                        if(method_exists($bimpObject, 'onValidate'))
+                            $errors = $bimpObject->onValidate($warnings);
+                        break;
 
-                case 'UNVALIDATE':
-                    $errors = $bimpObject->onUnvalidate($warnings);
-                    break;
+                    case 'UNVALIDATE':
+                        if(method_exists($bimpObject, 'onUnvalidate'))
+                            $errors = $bimpObject->onUnvalidate($warnings);
+                        break;
 
-                case 'DELETE':
-                    $errors = $bimpObject->onDelete($warnings);
-                    break;
+                    case 'DELETE':
+                        if(method_exists($bimpObject, 'onDelete'))
+                            $errors = $bimpObject->onDelete($warnings);
+                        break;
+                }
+
+                if (count($warnings)) {
+                    setEventMessages(BimpTools::getMsgFromArray($warnings), null, 'warnings');
+                }
+
+                if (count($errors)) {
+                    setEventMessages(BimpTools::getMsgFromArray($errors), null, 'errors');
+                    return -1;
+                }
             }
-
-            if (count($warnings)) {
-                setEventMessages(BimpTools::getMsgFromArray($warnings), null, 'warnings');
-            }
-            
-            if (count($errors)) {
-                setEventMessages(BimpTools::getMsgFromArray($errors), null, 'errors');
-                return -1;
-            }
+        } elseif (count($warnings)) {
+            setEventMessages(BimpTools::getMsgFromArray($warnings), null, 'warnings');
+            return 0;
         }
 
         if ($action == 'PAYMENT_CUSTOMER_DELETE') {
@@ -51,7 +60,8 @@ class InterfacePostProcess extends BimpCommTriggers
             if (BimpObject::objectLoaded($bimp_object)) {
                 $errors = $bimp_object->onDelete();
                 if (count($errors)) {
-                    setEventMessages(BimpTools::getMsgFromArray($errors), null, 'errors');
+                    $this->errors= $errors;
+//                    setEventMessages(BimpTools::getMsgFromArray($errors), null, 'errors');
                     return -1;
                 }
             }
